+++ crates/miden-protocol/src/account/interface/mod.rs
@@ -1,2 +1,5 @@
+mod code_interface;
+pub use code_interface::AccountCodeInterface;
+
 mod name;
 pub use name::AccountComponentName;